# 0-allocation mutable rem